==> 02_FE_modif_dataset.ipynb <==
#Importation des bibliothèques
import pandas as pd


#Ouverture du fichier 
df = pd.read_csv("ETL_csv.csv")

#df = df[df['isFraud'] == 1]

df.columns = ['id_trans', 'etape', 'type', 'montant', 'id_orig', 'old_solde_exp',
              'new_solde_exp', 'id_dest', 'old_solde_dest', 'new_solde_dest',
              'fraude']

# Filtrer les types qu'on souhaite retirer
types_to_remove = ['CASH_IN', 'DEBIT', 'PAYMENT']

# Créer un DataFrame avec uniquement les lignes à retirer
df_to_remove = df[df['type'].isin(types_to_remove)]

# Vérifiez combien de lignes peuvent être retirées
print(f"Taille du DataFrame à retirer : {df_to_remove.shape[0]}")

# Si le DataFrame à retirer a plus de 500 000 lignes, échantillonnez 500 000 lignes aléatoirement
if df_to_remove.shape[0] > 500000:
    df_to_remove = df_to_remove.sample(n=500000, random_state=1)  # random_state pour la reproductibilité

# Retirer les lignes sélectionnées du DataFrame original
df_final = df.drop(df_to_remove.index)

# Vérifiez la taille finale du dataset
print(f"Taille finale du dataset : {df_final.shape[0]}")

# Sauvegarder le nouveau dataset
df_final.to_csv('dataset_modifie.csv', index=False)  # Remplacez par le chemin de votre fichier de sortie

#Distribution des classes 
df_final['fraude'].value_counts()
df_final['fraude'].value_counts(normalize = True)
